+++ plugin/evm/config/config.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/ava-labs/avalanchego/utils/constants"
 	"github.com/ava-labs/avalanchego/vms/components/gas"
-	"github.com/ava-labs/coreth/plugin/evm/upgrade/etna"
 	"github.com/ava-labs/coreth/utils"
 	"github.com/ava-labs/libevm/common"
 	"github.com/ava-labs/libevm/common/hexutil"
@@ -68,7 +67,6 @@ const (
 	// Price Option Defaults
 	defaultPriceOptionSlowFeePercentage = uint64(95)
 	defaultPriceOptionFastFeePercentage = uint64(105)
-	defaultPriceOptionMaxBaseFee        = uint64(100 * utils.GWei)
 	defaultPriceOptionMaxTip            = uint64(20 * utils.GWei)
 )
 
@@ -152,7 +150,6 @@ type Config struct {
 	// Price Option Settings
 	PriceOptionSlowFeePercentage uint64 `json:"price-options-slow-fee-percentage"`
 	PriceOptionFastFeePercentage uint64 `json:"price-options-fast-fee-percentage"`
-	PriceOptionMaxBaseFee        uint64 `json:"price-options-max-base-fee"`
 	PriceOptionMaxTip            uint64 `json:"price-options-max-tip"`
 
 	TxPoolPriceLimit   uint64   `json:"tx-pool-price-limit"`
@@ -318,7 +315,6 @@ func (c *Config) SetDefaults(txPoolConfig TxPoolConfig) {
 	// Price Option Settings
 	c.PriceOptionSlowFeePercentage = defaultPriceOptionSlowFeePercentage
 	c.PriceOptionFastFeePercentage = defaultPriceOptionFastFeePercentage
-	c.PriceOptionMaxBaseFee = defaultPriceOptionMaxBaseFee
 	c.PriceOptionMaxTip = defaultPriceOptionMaxTip
 }
 
@@ -370,10 +366,6 @@ func (c *Config) Validate(networkID uint32) error {
 
 	if c.PushGossipPercentStake < 0 || c.PushGossipPercentStake > 1 {
 		return fmt.Errorf("push-gossip-percent-stake is %f but must be in the range [0, 1]", c.PushGossipPercentStake)
-	}
-
-	if c.PriceOptionMaxBaseFee < etna.MinBaseFee {
-		return fmt.Errorf("max base fee %d is less than the minimum base fee %d", c.PriceOptionMaxBaseFee, etna.MinBaseFee)
 	}
 	return nil
 }